+++ packages/ionic-react/src/schematics/init/schematic.spec.ts
@@ -1,3 +1,4 @@
+import * as ngSchematics from '@angular-devkit/schematics';
 import { Tree } from '@angular-devkit/schematics';
 import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
 import { readJsonInTree } from '@nrwl/workspace';
@@ -10,6 +11,11 @@ describe('init', () => {
   const testRunner = new SchematicTestRunner(
     '@nxtend/ionic-react',
     join(__dirname, '../../../collection.json')
+  );
+
+  testRunner.registerCollection(
+    '@nxtend/capacitor',
+    join(__dirname, '../../../../capacitor/collection.json')
   );
 
   beforeEach(() => {
@@ -57,6 +63,19 @@ describe('init', () => {
     expect(packageJson.dependencies['react-dom']).toBeDefined();
   });
 
+  it('should add and initialize nxtend Capacitor plugin', async () => {
+    const result = await testRunner
+      .runSchematicAsync('init', { capacitor: true }, appTree)
+      .toPromise();
+    const packageJson = readJsonInTree(result, 'package.json');
+
+    expect(packageJson.devDependencies['@nxtend/capacitor']).toBeDefined();
+    expect(packageJson.devDependencies['@nxtend/capacitor']).not.toEqual('*');
+
+    expect(packageJson.dependencies['@capacitor/core']).toBeDefined();
+    expect(packageJson.devDependencies['@capacitor/cli']).toBeDefined();
+  });
+
   it('should throw an error if Nrwl Workspace plugin is not installed', async () => {
     appTree.overwrite(
       'package.json',
@@ -80,5 +99,23 @@ describe('init', () => {
     const workspaceJson = readJsonInTree(result, 'workspace.json');
 
     expect(workspaceJson.cli.defaultCollection).toEqual('@nxtend/ionic-react');
+  });
+
+  describe('external schematics', () => {
+    it('should call the @nxtend/capacitor:init schematic', async () => {
+      const externalSchematicSpy = jest.spyOn(
+        ngSchematics,
+        'externalSchematic'
+      );
+      await testRunner
+        .runSchematicAsync('init', { capacitor: true }, appTree)
+        .toPromise();
+
+      expect(externalSchematicSpy).toBeCalledWith(
+        '@nxtend/capacitor',
+        'init',
+        expect.objectContaining({})
+      );
+    });
   });
 });