+++ packages/ionic-react/src/schematics/init/schematic.ts
@@ -1,6 +1,6 @@
-import { chain, noop, Rule } from '@angular-devkit/schematics';
-import { addPackageWithInit } from '@nrwl/workspace';
+import { chain, Rule } from '@angular-devkit/schematics';
 import { setDefaultCollection } from '@nrwl/workspace/src/utils/rules/workspace';
+import { addCapacitorPlugin } from './lib/add-capacitor-plugin';
 import { addDependencies } from './lib/add-dependencies';
 import { addReactPlugin } from './lib/add-react-plugin';
 import { InitSchematicSchema } from './schema';
@@ -9,7 +9,7 @@ export default function (options: InitSchematicSchema): Rule {
   return chain([
     setDefaultCollection('@nxtend/ionic-react'),
     addReactPlugin(),
-    options.capacitor ? addPackageWithInit('@nxtend/capacitor') : noop(),
+    addCapacitorPlugin(options),
     addDependencies(),
   ]);
 }